+++ manifest.json
@@ -10,10 +10,17 @@
   },
   "dependencies": {
     "vtex.search-resolver": "1.x",
-    "vtex.checkout-graphql": "0.x"
+    "vtex.checkout-graphql": "0.x",
+    "vtex.admin-cms-graphql": "0.x"
   },
   "credentialType": "absolute",
   "policies": [
+    {
+      "name": "POWER_USER_DS"
+    },
+    {
+      "name": "vtex.admin-cms-graphql:graphql-api"
+    },
     {
       "name": "graphql-query"
     },
