+++ node/package.json
@@ -15,10 +15,11 @@
     "isomorphic-unfetch": "^3.1.0"
   },
   "devDependencies": {
-    "@vtex/api": "6.37.0",
+    "@vtex/api": "6.37.1",
     "typescript": "3.9.7",
-    "vtex.checkout-graphql": "http://vtex.vtexassets.com/_v/public/typings/v1/vtex.checkout-graphql@0.46.0/public/@types/vtex.checkout-graphql",
-    "vtex.graphql-gateway": "https://gimenes--marinbrasil.myvtex.com/_v/private/typings/linked/v1/vtex.graphql-gateway@0.4.0+build1603980034/public/_types/react",
-    "vtex.search-resolver": "http://vtex.vtexassets.com/_v/public/typings/v1/vtex.search-resolver@1.22.3/public/_types/react"
+    "vtex.admin-cms-graphql": "https://gimenes--marinbrasil.myvtex.com/_v/private/typings/linked/v1/vtex.admin-cms-graphql@0.16.0-beta.0+build1609972022/public/@types/vtex.admin-cms-graphql",
+    "vtex.checkout-graphql": "http://vtex.vtexassets.com/_v/public/typings/v1/vtex.checkout-graphql@0.55.1/public/@types/vtex.checkout-graphql",
+    "vtex.graphql-gateway": "https://gimenes--marinbrasil.myvtex.com/_v/private/typings/linked/v1/vtex.graphql-gateway@0.7.0+build1609970821/public/_types/react",
+    "vtex.search-resolver": "http://vtex.vtexassets.com/_v/public/typings/v1/vtex.search-resolver@1.29.4/public/_types/react"
   }
 }
